% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearn_train_test_split.R
\name{sklearn_train_test_split}
\alias{sklearn_train_test_split}
\title{Sklearn train_test_split wrapper}
\usage{
sklearn_train_test_split(
  dataset,
  target_col,
  split,
  seed = NULL,
  return_only_index = FALSE
)
}
\arguments{
\item{dataset}{A data.table object. The dataset used for training.}

\item{target_col}{A character string. The name of the target column.}

\item{split}{A numeric. Ratio to split `dataset` into training set and
validation set (default: 0.7). The allowed value range is
0 < validation_split < 1.}

\item{seed}{A integer (default: NULL). Please use this argument in
order to generate reproducible results.}

\item{return_only_index}{A logical (default: FALSE). If FALSE,, the return
value is a list containing the split data.tables. If TRUE, only the
sampled row numbers are returned.}
}
\description{
A function to split a dataset into training and validation set,
  stratified by the target column.
}
\references{
\url{https://scikit-learn.org/stable/modules/generated/sklearn.
  model_selection.train_test_split.html}
}
